Creates CookieTablePanel consisting of Table and Details component
