+++ packages/extension/src/view/devtools/components/tabContent/cookies/components/cookieTablePanel/cookieDetails/details.tsx
@@ -22,7 +22,7 @@ import React, { useState } from 'react';
 /**
  * Internal dependencies.
  */
-import type { CookieTableData } from '../../../../../stateProviders/syncCookieStore';
+import type { CookieTableData } from '../../../../../../stateProviders/syncCookieStore';
 
 export interface DetailsProps {
   selectedCookie: CookieTableData;
