+++ packages/extension/src/view/devtools/components/tabContent/cookies/components/cookieTablePanel/cookieDetails/stories/cookieDetails.stories.ts
@@ -21,8 +21,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 /**
  * Internal dependencies.
  */
-
-import { CookieDetails } from '../..';
+import CookieDetails from '..';
 
 const meta = {
   title: 'Extension/DevTools/CookiesPanel/CookieDetails',
