+++ packages/extension/src/view/devtools/components/tabContent/cookies/components/cookieTablePanel/cookieTable/index.tsx
@@ -17,7 +17,7 @@
 /**
  * External dependencies.
  */
-import React, { useEffect, useMemo } from 'react';
+import React, { useMemo } from 'react';
 import {
   useReactTable,
   getCoreRowModel,
@@ -28,15 +28,13 @@ import {
 /**
  * Internal dependencies.
  */
-import Table from '../../../../table';
-import { useContentPanelStore } from '../../../../../stateProviders/contentPanelStore';
-import CheckIcon from '../../../../../../../../icons/check.svg';
-import type { CookieTableData } from '../../../../../stateProviders/syncCookieStore';
+import Table from '../../../../../table';
+import { useContentPanelStore } from '../../../../../../stateProviders/contentPanelStore';
+import CheckIcon from '../../../../../../../../../icons/check.svg';
+import type { CookieTableData } from '../../../../../../stateProviders/syncCookieStore';
 
 export interface CookieTableProps {
   cookies: CookieTableData[];
-  selectedFrame: string;
-  tabUrl: string | null;
 }
 
 const tableColumns: ColumnDef<CookieTableData>[] = [
@@ -146,11 +144,7 @@ const tableColumns: ColumnDef<CookieTableData>[] = [
   },
 ];
 
-const CookieTable = ({
-  cookies: data,
-  selectedFrame,
-  tabUrl,
-}: CookieTableProps) => {
+const CookieTable = ({ cookies: data }: CookieTableProps) => {
   const {
     selectedCookie,
     setSelectedCookie,
@@ -162,10 +156,6 @@ const CookieTable = ({
     tableColumnSize: state.tableColumnSize,
     tableContainerRef: state.tableContainerRef,
   }));
-
-  useEffect(() => {
-    setSelectedCookie(null);
-  }, [selectedFrame, tabUrl, setSelectedCookie]);
 
   const columns = useMemo(
     () =>
