+++ packages/extension/src/view/devtools/components/tabContent/cookies/components/cookieTablePanel/index.tsx
@@ -1,0 +1,77 @@
+/*
+ * Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * External dependencies.
+ */
+import React, { useEffect } from 'react';
+import { Resizable } from 're-resizable';
+
+/**
+ * Internal dependencies.
+ */
+import CookieTable from './cookieTable';
+import CookieDetails from './cookieDetails';
+import type { CookieTableData } from '../../../../../stateProviders/syncCookieStore';
+import { useContentPanelStore } from '../../../../../stateProviders/contentPanelStore';
+
+interface CookieTablePanelProps {
+  cookies: CookieTableData[];
+  selectedFrame: string | null;
+  tabUrl: string | null;
+}
+
+const CookieTablePanel = ({
+  cookies,
+  selectedFrame,
+  tabUrl,
+}: CookieTablePanelProps) => {
+  const { setSelectedCookie } = useContentPanelStore(({ actions }) => ({
+    setSelectedCookie: actions.setSelectedCookie,
+  }));
+
+  useEffect(() => {
+    setSelectedCookie(null);
+  }, [setSelectedCookie, tabUrl, selectedFrame]);
+
+  return (
+    selectedFrame && (
+      <div className="h-full flex flex-col">
+        <Resizable
+          defaultSize={{
+            width: '100%',
+            height: '80%',
+          }}
+          minHeight="6%"
+          maxHeight="98%"
+          enable={{
+            top: false,
+            right: false,
+            bottom: true,
+            left: false,
+          }}
+        >
+          <CookieTable cookies={cookies} />
+        </Resizable>
+        <div className="w-full h-full bg-white border-t-2 border-gray-300 shadow overflow-auto">
+          <CookieDetails />
+        </div>
+      </div>
+    )
+  );
+};
+
+export default CookieTablePanel;
